Collect code coverage as part of CI pipeline
With this change we collect code coverage information as part of the
GitLab CI pipeline. We use kcov which is able to gather such metrics
from a binary containing debug symbols, i.e., without recompilation to
provide additional instrumentation information.
The coverage report (in the form of web page detailing individual files
and lines of code) is included as an artifact to the pipeline run and
can be downloaded on demand. Unfortunately, it is not completely
straight forward to access it. In general it can be accessed at:

  https://d-e-s-o.gitlab.io/-/waitfor/-/jobs/<job-id>/artifacts/kcov/kcov-merged/index.html

Note that there are alternatives to this approach with online services
such as coveralls.io and codecov.io. However, ultimately those were
found to be either not working or insufficient. In particular,
codecov.io does not seem to provide line coverage information in the UI
itself, rendering the whole exercise somewhat useless. With coveralls.io
we have such information in principle and kcov is able to automatically
upload a report to it. However, kcov does not have any information about
the Git SHA-1 (or even the repository) used, as it just sees the binary.
The end result being that coveralls.io is not able to display any source
code information for a run. This could probably be fixed by issuing a
POST to their API manually (provided the GitLab CI environment contains
the necessary information), but ultimately that did not seem worth the
effort over what we have created here and now.